Frontend 5 (#47)
* fix some styling in delete confirmation modal

* Fix for Pyciu

* U mnie roomy dzialaja

+++ src/app/menu/menu-dashboard/delete-confirm-modal/delete-confirm-modal.component.ts
@@ -4,6 +4,7 @@ import { faTimes, faTimesCircle } from '@fortawesome/free-solid-svg-icons';
 import { RoomsService } from 'src/shared/services/rooms.service';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { LoginService } from 'src/shared/services/login.service';
 
 @Component({
   selector: 'app-delete-confirm-modal',
@@ -16,13 +17,16 @@ export class DeleteConfirmModalComponent {
 
   constructor(
     private roomsService: RoomsService,
+    private loginService: LoginService,
     public dialog: MatDialog,
     private router: Router,
     @Inject(MAT_DIALOG_DATA) public data: any
   ) {}
 
   onDeleteRoom(): void {
-    this.roomsService.deleteRoom(this.data.currentRoom._id);
+    console.log(this.loginService.getCurrentUser().username);
+    if (this.data.isAuthor) this.roomsService.deleteRoom(this.data.currentRoom._id);
+    else this.roomsService.leaveRoom(this.data.currentRoom.name, this.loginService.getCurrentUser().username);
     this.router.navigate(['/rooms']);
     this.closeModal();
   }
